requirements.txt
asgiref==3.2.7
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
mysqlclient==1.4.6
phonenumbers==8.12.2
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
